requirements.txt
 
# Make sure httpx is included:
httpx>=0.24.0


tiktoken==0.5.2


pytz==2023.3


# Core dependencies
pydantic-settings==2.7.0
pydantic_core==2.27.2
typing_extensions==4.12.2
annotated-types==0.7.0

# Database dependencies
alembic==1.14.0

# FastAPI and related
starlette==0.41.3
python-multipart==0.0.20
anyio==4.7.0
sniffio==1.3.1
h11==0.14.0

# HTTP and networking
httpcore==1.0.7
httpx==0.28.1
aiosmtplib==3.0.2
fastapi-mail==1.4.2

# Security and encryption
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
cryptography==44.0.0
cffi==1.17.1
pycparser==2.22
pyasn1==0.6.1
ecdsa==0.19.0
rsa==4.9
six==1.17.0

# Caching and rate limiting
redis==5.2.1
fastapi-limiter==0.1.6

# Email validation
email_validator==2.2.0
dnspython==2.7.0
idna==3.10

python-dotenv

# Templates and formatting
Jinja2==3.1.5
MarkupSafe==3.0.2
Mako==1.3.8

# Utilities
click==8.1.8
blinker==1.9.0
certifi==2024.12.14
distro==1.9.0
jiter==0.8.2
tqdm==4.67.1

# AI/ML
openai==1.58.1


# Payment
razorpay==1.3.0